fluu feture after production

+++ crm/src/app/controllers/user.js
@@ -12,6 +12,7 @@ const s3Service = require("../services/s3");
 const crypto = require("crypto");
 const Email = require("../services/email");
 const accountTableKey = require("../config/constants").accountTableKey;
+const { defultProfileImage } = require("../config/constants");
 /**
  * @desc    Login user
  * @route   POST /api/v1/users/login
@@ -75,12 +76,19 @@ exports.signup = async (req, res) => {
     req.body.email,
     req.body.phoneNumber,
   ];
+
   //  create account
   let result = await dbConnection.query(sqlQuery.CREATE_ACCOUNT, accountData);
   result = result.rows[0];
   // create client_tasks_state_repositry
   await dbConnection.query(sqlQuery.CREATE_TASKS_STATE_REPOSITORY, [
     result.user_id,
+  ]);
+
+  await dbConnection.query(sqlQuery.CREATE_ACCOUNT_PROFILE_IMAGE, [
+    result.user_id,
+    defultProfileImage.link,
+    defultProfileImage.s3_key,
   ]);
 
   res.status(201).json({
@@ -120,6 +128,7 @@ exports.updateUser = async (req, res) => {
     updateCol[accountTableKey[key]] = "";
     updateDate.push(req.body[key]);
   });
+
 
   const result = await dbConnection.query(
     sqlQuery.UPDATE_ACCOUNT_DATA(req.user.id, "account", updateCol),
@@ -173,7 +182,6 @@ exports.updatePassword = async (req, res) => {
  * @access  Private
  */
 exports.uploadProfileImage = async (req, res) => {
-  let profile_image_already_exist = false;
   if (!req.file)
     return res
       .status(400)
@@ -184,45 +192,24 @@ exports.uploadProfileImage = async (req, res) => {
     [req.user.id]
   );
 
-  if (userImage.rows[0]) {
+  if (
+    userImage.rows[0] &&
+    userImage.rows[0].s3_key != defultProfileImage.s3_key
+  )
     await s3Service.deleteOne(userImage.rows[0]);
-    profile_image_already_exist = true;
-  }
+
   const imageData = await s3Service.uploadFile(req.file);
 
-  if (profile_image_already_exist)
-    await dbConnection.query(sqlQuery.UPDATE_ACCOUNT_PROFILE_IMAGE, [
-      imageData.fileLink,
-      imageData.s3_key,
-      req.user.id,
-    ]);
-  else
-    await dbConnection.query(sqlQuery.CREATE_ACCOUNT_PROFILE_IMAGE, [
-      req.user.id,
-      imageData.fileLink,
-      imageData.s3_key,
-    ]);
+  await dbConnection.query(sqlQuery.UPDATE_ACCOUNT_PROFILE_IMAGE, [
+    imageData.fileLink,
+    imageData.s3_key,
+    req.user.id,
+  ]);
 
   res.status(200).json({
     message: "Successful upload",
     link: imageData.fileLink,
   });
-};
-/**
- * @desc    Get  profile image
- * @route   GET  /api/v1/users/profile/image
- * @access  Private
- */
-exports.getProfileImage = async (req, res) => {
-  const imageData = await dbConnection.query(
-    sqlQuery.GET_ACCOUNT_PROFILE_IMAGE,
-    [req.user.id]
-  );
-
-  if (!imageData.rows.length)
-    return res.status(400).json({ message: "image not uploaded" });
-
-  res.status(200).json({ imageLink: imageData.rows[0].link });
 };
 
 //  Forgot Password
@@ -327,4 +314,91 @@ exports.resetPassword = async (req, res) => {
   ]);
 
   res.status(200).json({ message: "Your password has been updated." });
+};
+
+//      just for super admin
+
+/**
+ * @desc    Delete user
+ * @route   POST /api/v1/super/user/:userId
+ * @access  Private/superAdmin
+ */
+exports.delete = async (req, res) => {
+  const userId = req.params.userId;
+  try {
+    await dbConnection.query("BEGIN;");
+    // delete folders and files
+    // --------------------------------------------------------------------
+    let foldersIds = await dbConnection.query(sqlQuery.GET_ALL_FOLDERS_IDS, [
+      userId,
+    ]);
+    foldersIds = foldersIds.rows;
+
+    // delete files
+    for (let i in foldersIds) {
+      let filesKeys = await dbConnection.query(sqlQuery.GET_FILES_S3_KEY, [
+        foldersIds[i].folder_id,
+      ]);
+      filesKeys = filesKeys.rows;
+      await s3Service.deleteManyFiles(filesKeys);
+      await dbConnection.query(sqlQuery.DELETE_ALL_FILES_BY_FOLDER_ID, [
+        foldersIds[i].folder_id,
+      ]);
+    }
+
+    // delete folders
+    await dbConnection.query(sqlQuery.DELETE_ALL_FOLDERS_BY_USER_ID, [userId]);
+    // --------------------------------------------------------------------
+    // delete projects and tasks
+    let projectsIds = await dbConnection.query(sqlQuery.GET_ALL_PROJECTS_IDS, [
+      userId,
+    ]);
+    projectsIds = projectsIds.rows;
+    // delete tasks
+    for (let i in projectsIds) {
+      await dbConnection.query(sqlQuery.DELETE_ALL_TASKS_FOR_PROJECT, [
+        projectsIds[i].pr_id,
+      ]);
+      // delete project worhing period
+      await dbConnection.query(
+        sqlQuery.DELETE_ALL_PROJECT_WORKING_PERIOD_BY_PROJECT_ID,
+        [projectsIds[i].pr_id]
+      );
+    }
+
+    // delete project
+    await dbConnection.query(sqlQuery.DELETE_ALL_PROJECTS_BY_USER_ID, [userId]);
+    // --------------------------------------------------------------------
+    // dekete user_tasks_state
+    await dbConnection.query(sqlQuery.DELETE_CLIENT_TASKS_STATE_BY_USER_ID, [
+      userId,
+    ]);
+    // check if profile image is defult
+    let profileImageS3Key = await dbConnection.query(
+      sqlQuery.GET_PROFILE_IMAGE_S3_KEY,
+      [userId]
+    );
+
+    profileImageS3Key = profileImageS3Key.rows[0].s3_key;
+    // delete user profile image
+    if (profileImageS3Key != defultProfileImage.s3_key)
+      await s3Service.deleteOne({ s3_key: profileImageS3Key });
+    await dbConnection.query(sqlQuery.DELETE_PROFILE_IMAGE_BY_USER_ID, [
+      userId,
+    ]);
+    // delete user account
+    await dbConnection.query(sqlQuery.DELETE_USER_ACCOUNT_BY_USER_ID, [userId]);
+
+    await dbConnection.query("COMMIT;");
+  } catch (err) {
+    console.log(err);
+    await dbConnection.query("ROLLBACK;");
+    return res.status(400).json({
+      message: "No valid entry found for provided ID",
+    });
+  }
+
+  res.status(200).json({
+    message: "User deleted",
+  });
 };
